login

+++ src/main/java/br/com/locacao/repository/ImovelRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.locacao.entity.Imovel;
 
-public interface ImovelRepository extends JpaRepository<Imovel, Integer> {
+public interface ImovelRepository extends JpaRepository<Imovel, Long> {
 }
-
